Fixed null mBgReading but, and cleand up code.
Was not testing for a null mBgReading object which happens when sensor is
stopped.  This crashes xDrip.

Set a single no_signal boolean for use throughout the class, that is set
once IF mBgReading is NOT null.  Allows for correct indication/return of
values when a packet has not been captured.

TODO:  Need to make the xBridge Battery return some form of indication
that this figure may be stale.  Right now it stays at the last value.  Not
sure that setting this back to 0 is the correct way to deal with it.